properly handle missing deps_ins
This was pretty subtle!  Superficially what I fixed here was that it's
legal for a deps_in to be missing, because of this scenario:
1) build foo.o, discover via deps it uses bar.h
2) edit foo.cc, remove the bar.h include, and delete bar.h
3) build foo.o -- needs to be ok that bar.h is missing

However, to get the above correct, I needed to revisit how I thought about
dependencies.  We now make a distinction between:
- ordering_ins: the inputs that must be visited in build order before
  this target; includes order-only edges
- deps_ins: the inputs that are discovered after building (TODO need a better
  name here)
- the inputs to the hash: the inputs that, if changed, invalidate the output,
  which are non-order-only inputs and deps_ins
- dirtying_ins: the same as above, excluding deps_ins; these are the inputs
  that we must stat() and find that they are present when checking a target

In all, this even discovered a build bug in the Figma build (a missing edge!)
which was triggering an n2 crash.  I have now promoted that to a user-facing
error.