redo how missing files work
This feels like a pretty big change.  I added doc/design_notes.md as
notes about what I concluded about missing files.  To summarize, it's
legal for any build to not write its outputs, but that means any
downstream build is responsible for possibly erroring if an input is
missing.