clean up handling of missing files
Move around some logic to separate out "check the state of all the files"
from "decide whether we need to run the build", so that the handling of
phoniness is clearer.  This fixes a bug where we'd still sometimes attempt
to "build" a phony build.

This also makes us bail out on stat()ing files as soon as we see any
of them missing, which is especially the common case in fresh builds.