build build.ninja if possible before running build
This was surprisingly subtle and it's still not quite right, in that
we cannot properly tell when you've generated build.ninja outside of
the build system.

The key here is that we want to load all the graph etc state,
attempt to build build.ninja, and if it's up to date, we want to
still use the loaded state to build the rest.  And then if it gets
rewritten we actually have to reload everything because it's likely
the graph changed.