refactor progress around task completion (#32)
Fixes a TODO around the progress API in the presence of keep-going flag.
Instead of two separate calls for task state change and task completion,
there's now one shared call for task start and finish, and it takes the
full TaskResult struct as a param.  This lets progress see the 'interrupted'
state of the task.