drop a needless use of HashSet for ready builds
We coincidentally at work had a problem that is structurally similar to
how n2 needs to manage a set of tasks in various states and compute different
views onto those states.

I was telling a coworker about how I went in circles around the right way
to represent this, like having a HashSet per state or not, and how I was
pretty happy about how the code ended up.  I went to look at it and realized
that this no longer even needs to be a HashSet.

This was a HashSet here and in Ninja because we need to be able to do two
things quickly:
1. given a BuildId, answer "is it ready?"
2. given free CPU resources, grab a ready build to work on it

But we now store a state per BuildId separately, handling #1, which means
that to provide behavior #2 we can just use a deque, which makes the
implementation of `pop_ready` now trivial.