don't swallow periods in variable names
Reverts 73704cd3ea35dec35fa915f39b9de5e9632c2d3e.
I'm not sure why we did that change, but looking at the Ninja source it
looks like there is "varname" (used in ${...}) and "simple_varname" (which
is what this code corresponds to) and the former allows '.' and the latter
doesn't.

Fixes #52.