avoid path canonicalization in graph API
Rearrange the API in the Graph type: it now expects the caller to have
canonicalized the path.  This simplifies the API by making it more
generic (perf-insensitive callers no longer need to provide a mutable buffer)
and lets us avoid unneeded canonicalization in the db loading path.

One way to see the layering here now is to note that the canonicalization
functions now are only called by load.rs and work.rs, both of which are
dealing with user-provided paths (.ninja files and command line, respectively).