use thread to centralize debouncing
Any time we want to update the fancy progress, instead set a flag that
a thread picks up to update after the fact.

We had an attempt at similar debouncing logic before that didn't work
reliably in the case where there's a status update immediately followed
by a long delay.  This new approach instead makes it so work.run() is
wholly synchronous, rather than attempting to flush/sleep, and it's up
to the progress reporter to do all the timing-related tricks.