rearrange display logic for smoother rendering
Before the code was like:
  loop { delay; wait for event; print }
This had a problem: when we are running a build task we're
sitting in the "wait" state.  When the task finished we'd immediately
print, so we'd print we were running n-1 of the possibly running tasks.
We'd kick off another task immediately but we'd be in the delay phase
at that point, so it'd be a while before the status updated.

The new logic is instead
  loop { wait for event; delay; print }
so it shows 100% utilization in most cases.