revisit how missing files work
To fix #69, I revisited how missing files work in n2 entirely.
This change includes an update to the design notes.  The summary is that
we no longer complain about a missing file as long as it's covered by the
build graph.

Interestingly, this means there is no part of n2 that cares specifically
about order-only inputs.  For build ordering purposes we use all inputs
(including order-only) to arrange the task order, but otherwise we never
consider the on-disk state of the files produced by an order-only dependency.

There is now a module in tests/e2e/missing.rs that cover some of the
missing file cases.

Fixes #69.