ensure state for all outputs of phony builds
If a build is missing its inputs or its outputs, we normally stop examining
disk state upon discovering the first missing file because we know we'll need
to execute it regardless.

But we don't execute phony builds, so they were skipping some logic that was
breaking an invariant: any step downstream of an output relies on that step
gathering the disk state (mtime) for that output.

For phony steps we expect outputs to be missing, and (due to working
around a Ninja bug) we also don't care about missing inputs(!).  Without
special casing, the "stop on first missing" file meant we wouldn't have
state for their outputs.

From a patch from liushuyu011@gmail.com, from #84.