stat outputs of phony rules
CMake has files that are outputs of a phony rule but still used as inputs.
  build ...: ... some_file
  build some_file: phony ...

Before this change, n2 would just mark all phony outputs as missing, which
breaks the above.  This change instead makes phony a little closer to how
non-phony rules work which fixes #40 and which I hope is a benefit in general.

Fixes #40.