avoid clown shoes allocation of input files
Scanner expects its input to be nul terminated, which I implemented as
  buf = std::fs::read(file)
  buf.push(0)

However, std::fs::read carefully sizes its buffer to be exactly the file's size,
which means the push() must grow the buffer.  This means for e.g. a 40mb input
file, we'd read it into a 40mb buffer, then grow the buffer, copying it into an
80mb buffer, just to add one byte!

Fix this by using a buffer of the appropriate size.

I attempted to measure the perf impact here and it wasn't measureable.
Possibly this buffer growth really is that fast?  It seems it will at least
have memory impact at least.