fix another \r\n bug
The n2 handling of escapes around newlines is very unsatisfying and hacky.
The underlying issue is that I am not clear on the Ninja rules for when
an escaped newline is allowed.  One idea is that escapes only happen in file names
or on the right side of variable definitions, while another is that you can escape
a newline anywhere in the text where you have a newline that you want to skip.

In particular, consider
  build$
   foo: ...
is that legal because it's the token `build` followed by (escaped) whitespace
followed by a filename, or illegal because we expect a space after `build`?

In any case this change preserves the existing n2 behavior, just making it
consistent in a test we already had but for \r\n newlines.