fix 'explain' test on windows
Without this change, the file appears unmodified after the second write.  Perhaps
it's racing because it's too fast?  But it appears the underlying operation is
using Windows FILETIME which has 100ns resolution.

It's not clear why, but changing the content of the file in this manner fixes the
test, at least as evidenced by a few runs on GitHub CI.